move trafficshape/Conn to its own file
